Separated the dom from the logic
