+++ index.html
@@ -1,13 +1,13 @@
 <!DOCTYPE html>
 <html>
 
-<head>
-  <title>Tic Tac Toe</title>
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-  <link rel="stylesheet" href="./css/style.css" />
-</head>
+  <head>
+    <title>Tic Tac Toe</title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
+    <link rel="stylesheet" href="./css/style.css" />
+  </head>
 
-<body>
+  <body>
     <div class="title">
         <h2>Tic Tac Toe</h2>
     </div>
@@ -58,8 +58,8 @@
       </table>
     </div>
 
-
-    <script type="module" src="js/index.js"></script>
-</body>
+    <script src="js/index.js"></script>
+    <script src="js/dom.js"></script>
+  </body>
 
 </html>
